+++ ros-20190910/core/Models/GetStackGroupRequest.cs
@@ -17,6 +17,10 @@ namespace AlibabaCloud.SDK.ROS20190910.Models
         [Validation(Required=false)]
         public string StackGroupName { get; set; }
 
+        [NameInMap("StackGroupId")]
+        [Validation(Required=false)]
+        public string StackGroupId { get; set; }
+
     }
 
 }