+++ ros-20190910/core/Models/GetStackGroupResponseBody.cs
@@ -86,6 +86,9 @@ namespace AlibabaCloud.SDK.ROS20190910.Models
                 public int? DriftedStackInstancesCount { get; set; }
 
             }
+            [NameInMap("ResourceGroupId")]
+            [Validation(Required=false)]
+            public string ResourceGroupId { get; set; }
         };
 
     }