+++ ros-20190910/core/Models/GetStackRequest.cs
@@ -21,6 +21,14 @@ namespace AlibabaCloud.SDK.ROS20190910.Models
         [Validation(Required=false)]
         public string ClientToken { get; set; }
 
+        [NameInMap("OutputOption")]
+        [Validation(Required=false)]
+        public string OutputOption { get; set; }
+
+        [NameInMap("ShowResourceProgress")]
+        [Validation(Required=false)]
+        public string ShowResourceProgress { get; set; }
+
     }
 
 }