+++ ros-20190910/core/Models/ListStackGroupsResponseBody.cs
@@ -37,6 +37,10 @@ namespace AlibabaCloud.SDK.ROS20190910.Models
             [Validation(Required=false)]
             public string StackGroupName { get; set; }
 
+            [NameInMap("ResourceGroupId")]
+            [Validation(Required=false)]
+            public string ResourceGroupId { get; set; }
+
         }
 
         [NameInMap("TotalCount")]