+++ ros-20190910/core/Models/ListStacksResponseBody.cs
@@ -95,6 +95,10 @@ namespace AlibabaCloud.SDK.ROS20190910.Models
             [Validation(Required=false)]
             public string StackType { get; set; }
 
+            [NameInMap("ResourceGroupId")]
+            [Validation(Required=false)]
+            public string ResourceGroupId { get; set; }
+
         }
 
     }