+++ ros-20190910/core/Models/MoveResourceGroupResponseBody.cs
@@ -1,0 +1,21 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ROS20190910.Models
+{
+    public class MoveResourceGroupResponseBody : TeaModel {
+        /// <summary>
+        /// Id of the request
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}